Add something really fun with template with a Quang
